requirements.txt
absl-py==2.4.0
alembic==1.18.4
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
blosc2==4.1.2
cachetools==7.0.4
certifi==2026.2.25
charset-normalizer==3.4.5
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-core==0.6.0
cuda-pathfinder==1.4.1
cuda-python==12.9.5
cuda-toolkit==12.9.1
cudf-cu12==26.2.1
cupy-cuda12x==14.0.1
cycler==0.12.1
DateTime==6.0
debugpy==1.8.20
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
gast==0.7.0
google-pasta==0.2.0
greenlet==3.3.2
grpcio==1.78.0
h5py==3.14.0
idna==3.11
ipykernel==7.2.0
ipython==8.38.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
keras==3.12.1
kiwisolver==1.4.9
libclang==18.1.1
libcudf-cu12==26.2.1
libkvikio-cu12==26.2.0
librmm-cu12==26.2.0
lightning-utilities==0.15.3
llvmlite==0.44.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
mysql-connector-python==9.6.0
namex==0.1.0
narwhals==2.17.0
nbformat==5.10.4
ndindex==1.10.1
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numba-cuda==0.22.2
numexpr==2.14.1
numpy==2.2.6
openpyxl==3.1.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cccl-cu12==12.9.27
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-libnvcomp-cu12==5.1.0.21
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
nvtx==0.2.14
opt_einsum==3.4.0
optree==0.19.0
optuna==4.8.0
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
prompt_toolkit==3.0.52
protobuf==7.34.0
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==23.0.1
Pygments==2.19.2
pylibcudf-cu12==26.2.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyxlsb==1.0.10
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
rapids-logger==0.2.3
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rmm-cu12==26.2.0
rpds-py==0.30.0
ruff==0.15.6
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.48
stack-data==0.6.3
sympy==1.14.0
tables==3.10.1
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tomli==2.4.0
torch==2.10.0
torchmetrics==1.9.0
torchvision==0.25.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
triton==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
wrapt==2.1.2
xlrd==2.0.1
xgboost==3.2.0
zope.interface==8.2
